services/provisioner-api/pyproject.toml
[tool.poetry]
name = "provisioner-api"
version = "1.0.0"
description = "VM provisioning and lifecycle management for cloud gaming"
authors = ["Gamer Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
google-cloud-compute = "^1.30.0"
google-cloud-billing = "^1.12.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymongo = "^4.6.1"
geopy = "^2.4.1"
cryptography = "^41.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"